Add SBB, CMPXCHG, XCHG, ADC and fix memory operations
New instructions:
- SBB (subtract with borrow) reg-imm, reg-reg
- CMPXCHG 32-bit and 64-bit memory operations
- XCHG reg-reg, reg-mem, mem-reg variants
- ADC (add with carry) reg-imm
- CLD/STD/HLT instruction handlers
- PUSH [mem] support
- SUB [mem], reg including r8-r15 base
- ADD [mem], r8-r15 source registers
- OR [mem], imm read-modify-write

Bug fixes:
- CLD/STD: use 32-bit store for df_offset field
- REP string ops: use ldrsw for df_offset (32-bit signed)
- Memory R-M-W: save/restore guest address pattern

busybox echo now works on 64-bit emulation.

+++ kernel/calls.h
@@ -256,7 +256,8 @@ int_t sys_syslog(int_t type, addr_t buf_addr, int_t len);
 int_t sys_ipc(uint_t call, int_t first, int_t second, int_t third, addr_t ptr, int_t fifth);
 
 #ifdef ISH_GUEST_64BIT
-typedef int (*syscall_t)(qword_t, qword_t, qword_t, qword_t, qword_t, qword_t);
+// For 64-bit mode, syscalls like brk/mmap return 64-bit addresses
+typedef int64_t (*syscall_t)(qword_t, qword_t, qword_t, qword_t, qword_t, qword_t);
 #else
 typedef int (*syscall_t)(dword_t, dword_t, dword_t, dword_t, dword_t, dword_t);
 #endif
